==> 0_package_installer_script.R <==

if (!require("caret")) install.packages("caret")
if (!require("effects")) install.packages("effects")
if (!require("readr")) install.packages("readr")
if (!require("dplyr")) install.packages("dplyr")
if (!require("tidyr")) install.packages("tidyr")
if (!require("WriteXLS")) install.packages("WriteXLS")
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("reshape2")) install.packages("reshape2")
if (!require("car")) install.packages("car")
if (!require("glmnet")) install.packages("glmnet")
