==> 2_efv.R <==
library(caret)
library(effects)

library(readr)
library(dplyr)
library(tidyr)
library(WriteXLS)

library(ggplot2)
library(reshape2)



data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)




print(table(data_df$Species))
print(table(data_df$labels))




###############################################################
cor_matrix <- cor(data_df %>% select(-Species), use = "complete.obs")
cor_melted <- melt(cor_matrix)
cor_df <- as.data.frame(cor_matrix)
cor_df$Feature <- rownames(cor_df)
cor_df <- cor_df %>% select(Feature, everything())
WriteXLS::WriteXLS(cor_df, ExcelFileName = "Correlation_Matrix.xlsx")
###############################################################



#feature corrplot
###############################################################
data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)

get_lower_tri <- function(mat) {
  mat[upper.tri(mat)] <- NA
  return(mat)
}

cor_matrix <- cor(data_df %>% select(-Species), use = "complete.obs")
lower_tri <- get_lower_tri(cor_matrix)
lower_tri_melt <- melt(lower_tri, na.rm = TRUE)

dev.new()
ggplot(lower_tri_melt, aes(x = Var2, y = Var1, fill = value)) +
  geom_tile(color = "white") +
  geom_text(aes(label = round(value, 2)), color = "black", size = 3) +
  scale_fill_gradient2(low = "blue", high = "red", mid = "white",
                       midpoint = 0, limit = c(-1, 1), space = "Lab",
                       name = "Correlation") +
  scale_x_discrete(position = "top") +  # Move x-axis to the top
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 0, vjust = 0.5),
        axis.title.x = element_blank(),
        axis.title.y = element_blank()) +
  coord_fixed()
###############################################################






#boxplot
###############################################################
data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)
data_df$labels <- factor(data_df$labels)
set_order <- c("B", "G", "R", "RE", "NIR", "GLI", "ExG", "ExR", 
                   "ExGR", "RGI", "CIVE", "NGRDI", "VARI", "NDVI", "EVI",
                   "NDRE", "RECI", "GCI", "SR", "Ht")


data_long <- data_df %>%
  pivot_longer(cols = -c(Species, labels), 
               names_to = "Feature", 
               values_to = "Value")

data_long$Feature <- factor(data_long$Feature, levels = set_order)

labels_numeric <- as.numeric(as.character(data_df$labels))

correlations <- sapply(set_order, function(feat) {
  cor(data_df[[feat]], labels_numeric, use = "complete.obs")
})


new_labels <- sapply(set_order, function(x) {
  paste0(x, "\n(Corr: ", round(correlations[x], 2), ")")
})
names(new_labels) <- set_order


dev.new()
ggplot(data_long, aes(x = labels, y = Value, fill = labels)) +
  geom_boxplot() +
  facet_wrap(~ Feature, scales = "free_y", labeller = as_labeller(new_labels)) +
  labs(x = "Label (0 = WS, 1 = TA)", 
       y = "", 
       title = "") +
  theme_minimal() +
  theme(legend.position = "none",
        strip.text = element_text(size = 10))
###############################################################







#jitter
###############################################################
data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)
data_df$labels <- factor(data_df$labels)

set_order <- c("B", "G", "R", "RE", "NIR", "GLI", "ExG", "ExR", 
                   "ExGR", "RGI", "CIVE", "NGRDI", "VARI", "NDVI", "EVI",
                   "NDRE", "RECI", "GCI", "SR", "Ht")


data_long <- data_df %>%
  pivot_longer(cols = -c(Species, labels), 
               names_to = "Feature", 
               values_to = "Value")


data_long$Feature <- factor(data_long$Feature, levels = set_order)


labels_numeric <- as.numeric(as.character(data_df$labels))
correlations <- sapply(set_order, function(feat) {
  cor(data_df[[feat]], labels_numeric, use = "complete.obs")
})

new_labels <- sapply(set_order, function(x) {
  paste0(x, "\n(Corr: ", round(correlations[x], 2), ")")
})
names(new_labels) <- set_order


dev.new()
ggplot(data_long, aes(x = labels, y = Value, color = labels)) +
  geom_jitter(width = 0.2, size = 1.5, alpha = 0.6) +
  facet_wrap(~ Feature, scales = "free_y", labeller = as_labeller(new_labels)) +
  labs(x = "Label (0 = WS, 1 = TA)", 
       y = "", 
       title = "") +
  theme_minimal() +
  theme(legend.position = "none",
        strip.text = element_text(size = 10))
###############################################################







#scatter
###############################################################
data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)
data_df$labels <- factor(data_df$labels)


data_df$index <- 1:nrow(data_df)


set_order <- c("B", "G", "R", "RE", "NIR", "GLI", "ExG", "ExR", 
                   "ExGR", "RGI", "CIVE", "NGRDI", "VARI", "NDVI", "EVI",
                   "NDRE", "RECI", "GCI", "SR", "Ht")


data_long <- data_df %>%
  pivot_longer(cols = -c(Species, labels, index), 
               names_to = "Feature", 
               values_to = "Value")


data_long$Feature <- factor(data_long$Feature, levels = set_order)


labels_numeric <- as.numeric(as.character(data_df$labels))
correlations <- sapply(set_order, function(feat) {
  cor(data_df[[feat]], labels_numeric, use = "complete.obs")
})


new_labels <- sapply(set_order, function(x) {
  paste0(x, "\n(Corr: ", round(correlations[x], 2), ")")
})
names(new_labels) <- set_order


dev.new()
ggplot(data_long, aes(x = index, y = Value, color = labels)) +
  geom_point() +
  facet_wrap(~ Feature, scales = "free_y", labeller = as_labeller(new_labels)) +
  labs(x = "Label (0 = WS, 1 = TA)", y = "", 
       title = "") +
  theme_minimal() +
  theme(legend.position = "bottom",
        strip.text = element_text(size = 10))
###############################################################








#kernel
###############################################################
data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)
data_df$labels <- factor(data_df$labels)


set_order <- c("B", "G", "R", "RE", "NIR", "GLI", "ExG", "ExR", 
                   "ExGR", "RGI", "CIVE", "NGRDI", "VARI", "NDVI", "EVI",
                   "NDRE", "RECI", "GCI", "SR", "Ht")


data_long <- data_df %>%
  pivot_longer(cols = -c(Species, labels), 
               names_to = "Feature", 
               values_to = "Value")


data_long$Feature <- factor(data_long$Feature, levels = set_order)


labels_numeric <- as.numeric(as.character(data_df$labels))
correlations <- sapply(set_order, function(feat) {
  cor(data_df[[feat]], labels_numeric, use = "complete.obs")
})


new_labels <- sapply(set_order, function(x) {
  paste0(x, "\n(Corr: ", round(correlations[x], 2), ")")
})
names(new_labels) <- set_order


dev.new()
ggplot(data_long, aes(x = Value, fill = labels, color = labels)) +
  geom_density(alpha = 0.4) +  # Transparency to overlay
  facet_wrap(~ Feature, scales = "free", labeller = as_labeller(new_labels)) +
  labs(x = "Label (0 = WS, 1 = TA)", y = "", 
       title = "") +
  theme_minimal() +
  theme(legend.position = "bottom",
        strip.text = element_text(size = 10))
###############################################################











#tree height~feature corrs
###############################################################
data_df <- read.csv("0_WS_1_TA_data.csv", stringsAsFactors = FALSE)
data_df$labels <- factor(data_df$labels)


set_order <- c("B", "G", "R", "RE", "NIR", "GLI", "ExG", "ExR", 
                      "ExGR", "RGI", "CIVE", "NGRDI", "VARI", "NDVI", "EVI",
                      "NDRE", "RECI", "GCI", "SR")


data_long <- data_df %>%
  pivot_longer(cols = all_of(set_order), 
               names_to = "Feature", 
               values_to = "Value")


data_long$Feature <- factor(data_long$Feature, levels = set_order)


dev.new()
ggplot(data_long, aes(x = Value, y = Ht, color = labels)) +
  geom_point(alpha = 0.4) +
  geom_smooth(method = "lm", se = TRUE) +
  facet_wrap(~ Feature, scales = "free_x") +
  labs(x = "Label (0 = WS, 1 = TA)", y = "Height (Ht)", color = "Label",
       title = "") +
  theme_minimal() +
  theme(strip.text = element_text(size = 10),
        legend.position = "bottom")  
###############################################################
