==> 4_five_fold_binomial_functions_caller.R <==
source("4_five_fold_binomial_functions.R")





#formula <- labels ~ B + G + R + RE + NIR + GLI + ExG + ExR + ExGR +
#  RGI + CIVE + NGRDI + VARI + NDVI + EVI + NDRE + RECI + GCI + SR + Ht
#formula <- labels ~ VARI+GLI+B+NGRDI+NIR+NDRE+R+RECI+Ht+EVI+SR+GCI+RE
#formula <- labels ~ VARI + GLI + B + NGRDI + NIR
#formula <- labels ~ B + R + NIR + GLI + NGRDI + VARI + EVI + NDRE + RECI + Ht
#formula <- labels ~ NIR + NDVI
#formula <- labels ~ NDVI + NIR + RE + Ht


#formula <- labels ~ Ht*(NIR + NDVI)

formula <- labels ~ Ht*(B + R + NIR + GLI + NGRDI + VARI + EVI + NDRE + RECI + Ht)


#5 fold cross vlaidation caller
cv_results <- def.run.binomial.5.fold.cv(formula, "0_TA_1_WS_data.csv")


#confusion matrix visualiztion 
plot_confusion_local(cv_results$all_preds_factor_local, 
                     cv_results$all_actual_factor_local)

#ROC plot for Sup. fig 8
plot_roc_auc_local(cv_results$roc_list_local, cv_results$auc_vec_local)
